backend/requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiohttp-cache==4.0.1
aiohttp-devtools==1.1.2
aiosignal==1.3.1
alembic==1.13.2
annotated-types==0.7.0
anyio==4.4.0
asttokens==2.4.1
async-timeout==4.0.3
asyncpg==0.29.0
attrs==24.2.0
black==24.8.0
click==8.1.7
devtools==0.12.2
dnspython==2.6.1
email_validator==2.2.0
envparse==0.2.0
exceptiongroup==1.2.2
executing==2.0.1
frozenlist==1.4.1
greenlet==3.0.3
idna==3.8
Mako==1.3.5
MarkupSafe==2.1.5
multidict==6.0.5
mypy-extensions==1.0.0
packaging==24.1
passlib==1.7.4
pathspec==0.12.1
pika==1.3.2
platformdirs==4.3.6
psycopg2-binary==2.9.9
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
PyJWT==2.9.0
python-dotenv==1.0.1
redis==4.6.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.32
typing_extensions==4.12.2
watchfiles==0.23.0
yarl==1.9.4
